+++ lib/munola/railtie.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Munola
-  class Railtie < Rails::Railtie
-    generators { require_relative "../generators/munola/install/install_generator" }
-  end
-end